requirements.txt
web3==6.11.1
python-dotenv==1.0.0
requests==2.31.0
python-telegram-bot==20.7
aiohttp==3.9.1